+++ plugins/sudoers/ldap_conf.c
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2003-2018 Todd C. Miller <Todd.Miller@sudo.ws>
+ * Copyright (c) 2003-2019 Todd C. Miller <Todd.Miller@sudo.ws>
  *
  * This code is derived from software contributed by Aaron Spangler.
  *
@@ -199,7 +199,7 @@ sudo_ldap_conf_add_ports(void)
 
     hostbuf[0] = '\0';
     len = snprintf(defport, sizeof(defport), ":%d", ldap_conf.port);
-    if (len <= 0 || (size_t)len >= sizeof(defport)) {
+    if (len <= 0 || len >= (int)sizeof(defport)) {
 	sudo_warnx(U_("sudo_ldap_conf_add_ports: port too large"));
 	debug_return_bool(false);
     }
